// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/Enumerations.ts
export enum SoeEntityState { 
    Active = 0,
    Inactive = 1,
    Deleted = 2,
    Temporary = 3,
    Hidden = 4
}