// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/ProductDTOs.ts
// This file was generated by TypeWriter. See TypeWriter.Common.tst for reference.



    export interface IProductSmallDTO {
        productId: number;
        number: string;
        name: string;
        numberName: string;
    }

    

    

    export interface IProductPriceListDTO {
        priceListId: number;
        purchasePrice: number;
        price: number;
        startDate: Date;
        stopDate: Date;
    }

    

    

    export interface IProductComparisonDTO {
        purchasePrice: number;
        comparisonPrice: number;
        price: number;
        startDate: Date;
        stopDate: Date;
    }

    

    

    export interface IProductDTO {
        productUnitId?: number;
        productGroupId?: number;
        type: number;
        description: string;
        accountingPrio: string;
        created?: Date;
        createdBy: string;
        modified?: Date;
        modifiedBy: string;
        state: number;
        productUnitCode: string;
    }

    

    

    export interface IProductCleanupDTO {
        productId: number;
        productNumber: string;
        productName: string;
        isExternal: boolean;
        lastUsedDate: Date;
        isActive: boolean;
    }

    

    

    export interface IProductTimeCodeDTO {
        id: number;
        name: string;
        vatType?: TermGroup_InvoiceProductVatType;
        state: number;
    }

    

    






import { TermGroup_InvoiceProductVatType } from "../generated-interfaces/Enumerations";