// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/InvoiceProductDTOs.ts
// This file was generated by TypeWriter. See TypeWriter.Common.tst for reference.



    export interface IInvoiceProductSmallDTO {
        calculationType: number;
        productUnitId?: number;
        productGroupId?: number;
        guaranteePercentage?: number;
        useCalculatedCost?: boolean;
        purchasePrice: number;
    }

    

    

    export interface IInvoiceProductDTO {
        sysProductId?: number;
        sysPriceListHeadId?: number;
        vatType: TermGroup_InvoiceProductVatType;
        vatFree: boolean;
        ean: string;
        purchasePrice: number;
        sysWholesellerName: string;
        calculationType: TermGroup_InvoiceProductCalculationType;
        guaranteePercentage?: number;
        timeCodeId?: number;
        priceListOrigin: number;
        showDescriptionAsTextRow: boolean;
        showDescrAsTextRowOnPurchase: boolean;
        dontUseDiscountPercent: boolean;
        useCalculatedCost: boolean;
        vatCodeId?: number;
        householdDeductionType?: number;
        householdDeductionPercentage?: number;
        isStockProduct: boolean;
        weight?: number;
        intrastatCodeId?: number;
        sysCountryId?: number;
        defaultGrossMarginCalculationType?: number;
        isExternal?: boolean;
        salesPrice: number;
        isSupplementCharge: boolean;
        priceLists: IPriceListDTO[];
        categoryIds: number[];
        accountingSettings: IAccountingSettingsRowDTO[];
        sysProductType?: number;
    }

    

    

    export interface IInvoiceProductGridDTO {
        productId: number;
        number: string;
        numberSort: string;
        name: string;
        productGroupId?: number;
        productGroupCode: string;
        productGroupName: string;
        productCategories: string;
        sysProductId?: number;
        external: boolean;
        vatType: TermGroup_InvoiceProductVatType;
        state: SoeEntityState;
        eanCode: string;
        timeCodeId?: number;
        timeCodeName: string;
        isActive?: boolean;
    }

    

    

    export interface IProductRowsProductDTO {
        productId: number;
        sysProductId?: number;
        number: string;
        name: string;
        description: string;
        showDescriptionAsTextRow: boolean;
        showDescrAsTextRowOnPurchase: boolean;
        productUnitId?: number;
        productUnitCode: string;
        vatType: TermGroup_InvoiceProductVatType;
        calculationType: TermGroup_InvoiceProductCalculationType;
        guaranteePercentage?: number;
        sysWholesellerName: string;
        dontUseDiscountPercent: boolean;
        purchasePrice: number;
        salesPrice: number;
        vatCodeId?: number;
        weight?: number;
        isExternal: boolean;
        isStockProduct: boolean;
        isSupplementCharge: boolean;
        householdDeductionType?: number;
        householdDeductionPercentage?: number;
        isLiftProduct: boolean;
        isInactive: boolean;
        intrastatCodeId?: number;
        sysCountryId?: number;
        grossMarginCalculationType: TermGroup_GrossMarginCalculationType;
    }

    

    






import { SoeEntityState, TermGroup_InvoiceProductVatType, TermGroup_InvoiceProductCalculationType, TermGroup_GrossMarginCalculationType } from "../generated-interfaces/Enumerations";
import { IPriceListDTO } from "../generated-interfaces/PriceListDTOs";
import { IAccountingSettingsRowDTO } from "../generated-interfaces/SOECompModelDTOs";