// Soe.Angular.Spa/src/app/shared/models/generated-interfaces/PriceListTypeDTOs.ts
// This file was generated by TypeWriter. See TypeWriter.Common.tst for reference.



    export interface IPriceListTypeDTO {
        priceListTypeId: number;
        currencyId: number;
        name: string;
        description: string;
        discountPercent: number;
        inclusiveVat: boolean;
        isProjectPriceList: boolean;
        created?: Date;
        createdBy: string;
        modified?: Date;
        modifiedBy: string;
        state: SoeEntityState;
    }

    

    

    export interface IPriceListTypeGridDTO {
        priceListTypeId: number;
        currencyId: number;
        sysCurrencyId?: number;
        name: string;
        description: string;
        inclusiveVat: boolean;
        currency: string;
        isProjectPriceList: boolean;
    }

    

    






import { SoeEntityState } from "../generated-interfaces/Enumerations";